+++ aerosense/__init__.py
@@ -13,8 +13,8 @@ Usage:
 from typing import List
 
 # --- Versioning ---
-__version__: str = "v3.9.1"
-__release__: str = "2026-02-18"
+__version__: str = "v3.10.0"
+__release__: str = "2026-03-20"
 
 # --- Component Exports ---
 from .core.controller import Controller